requirements.txt
packaging
sentencepiece
peft==0.4.0
einops==0.4.1
fastapi==0.100.1
gradio==3.39.0
markdown2==2.4.10
# numpy==1.24.2
openai==0.27.8
opencv_python==4.8.0.74
Pillow==9.4.0
# pycocotools==2.0.6
ray==2.6.1
Requests==2.31.0
shortuuid==1.0.11
tqdm==4.64.1
transformers==4.31.0
uvicorn==0.23.2
# huggingface-hub==0.19.4
mmcv==2.0.0 -f https://download.openmmlab.com/mmcv/dist/cu117/torch2.0/index.html

pycocotools
deepspeed
tensorboard
huggingface_hub>=0.23,<0.27
scipy
torchvision
scikit-image